Struct holochain_container_api::config::Configuration

pub struct Configuration {
    pub agents: Vec<AgentConfiguration>,
    pub dnas: Vec<DnaConfiguration>,
    pub instances: Vec<InstanceConfiguration>,
    pub interfaces: Vec<InterfaceConfiguration>,
    pub bridges: Vec<Bridge>,
}

Main container configuration struct This is the root of the configuration tree / aggregates all other configuration aspects.

References between structs (instance configs pointing to the agent and DNA to be instantiated) are implemented via string IDs.

Fields
agents: Vec<AgentConfiguration>

List of Agents, this mainly means identities and their keys. Required.

dnas: Vec<DnaConfiguration>

List of DNAs, for each a path to the DNA file. Required.

instances: Vec<InstanceConfiguration>

List of instances, includes references to an agent and a DNA. Required.

interfaces: Vec<InterfaceConfiguration>

List of interfaces any UI can use to access zome functions. Optional.

bridges: Vec<Bridge>

List of bridges between instances. Optional.

Methods

impl Configuration

pub fn check_consistency(&self) -> Result<(), String>

This function basically checks if self is a semantically valid configuration. This mainly means checking for consistency between config structs that reference others.

pub fn agent_by_id(&self, id: &str) -> Option<AgentConfiguration>

Returns the agent configuration with the given ID if present

pub fn dna_by_id(&self, id: &str) -> Option<DnaConfiguration>

Returns the DNA configuration with the given ID if present

pub fn instance_by_id(&self, id: &str) -> Option<InstanceConfiguration>

Returns the instance configuration with the given ID if present

pub fn interface_by_id(&self, id: &str) -> Option<InterfaceConfiguration>

Returns the interface configuration with the given ID if present

pub fn instance_ids(&self) -> Vec<String>

Returns all defined instance IDs

pub fn instance_ids_sorted_by_bridge_dependencies(
&self
) -> Result<Vec<String>, HolochainError>

This function uses the petgraph crate to model the bridge connections in this config as a graph and then create a topological sorting of the nodes, which are instances. The sorting gets reversed to get those instances first that do NOT depend on others such that this ordering of instances can be used to spawn them and simultaneously create initialize the bridges and be able to assert that any callee already exists (which makes this task much easier).

pub fn bridge_dependencies(&self, caller_instance_id: String) -> Vec<Bridge>