Struct holochain_container_api::container::Container

pub struct Container { /* fields omitted */ }

Main representation of the container. Holds a HashMap of Holochain instances referenced by ID. A primary point in this struct is load_config(&mut self, config: &Configuration) -> Result<(), String> which takes a config::Configuration struct and tries to instantiate all configured instances. While doing so it has to load DNA files referenced in the configuration. In order to not bind this code to the assumption that there is a filesystem and also enable easier testing, a DnaLoader ()which is a closure that returns a Dna object for a given path string) has to be injected on creation.

Methods

impl Container

pub fn from_config(config: Configuration) -> Self

Creates a new instance with the default DnaLoader that actually loads files.

pub fn with_signal_channel(self, signal_tx: SyncSender<Signal>) -> Self

pub fn config(&self) -> Configuration

pub fn start_all_interfaces(&mut self)

pub fn start_interface_by_id(&mut self, id: String) -> Result<(), String>

pub fn start_all_instances(&mut self) -> Result<(), HolochainInstanceError>

Starts all instances

pub fn stop_all_instances(&mut self) -> Result<(), HolochainInstanceError>

Stops all instances

pub fn instances(&self) -> &InstanceMap

pub fn shutdown(&mut self) -> Result<(), HolochainInstanceError>

Stop and clear all instances

pub fn spawn_network(&mut self) -> Result<String, HolochainError>

pub fn load_config(&mut self) -> Result<(), String>

Tries to create all instances configured in the given Configuration object. Calls Configuration::check_consistency() first and clears self.instances. The first time we call this, we also initialize the container-wide config for use with all instances

@TODO: clean up the container creation process to prevent loading config before proper setup, especially regarding the signal handler. (see https://github.com/holochain/holochain-rust/issues/739)

pub fn instantiate_from_config(
&mut self,
id: &String,
config: &Configuration
) -> Result<Holochain, String>

Creates one specific Holochain instance from a given Configuration, id string and DnaLoader.